+++ theme_gen/css/__init__.py
@@ -1,0 +1,1 @@
+"""CSS generation for markdown preview styling."""